FIXED: Consolidated deserializer and converter
