+++ src/lib/kotlin/slatekit-tests/src/test/kotlin/test/common/DeserializerTests.kt
@@ -9,7 +9,7 @@ import slatekit.common.encrypt.EncDouble
 import slatekit.common.encrypt.EncInt
 import slatekit.common.encrypt.EncLong
 import slatekit.common.encrypt.EncString
-import slatekit.meta.Converter
+import slatekit.meta.Deserializer
 import java.time.LocalDate
 import java.time.LocalDateTime
 import java.time.LocalTime
@@ -36,8 +36,8 @@ class ConvertTests {
     fun test_basic_types(tstr:String, tbool:Boolean, tshort:Short, tint:Int, tlong:Long, tdoub:Double):Unit {}
     @Test fun can_parse_basictypes(){
         val test = """{ "tstr": "abc", "tbool": false, "tshort": 1, "tint": 12, "tlong": 123, "tdoub": 123.45 }"""
-        val converter = Converter()
-        val results = converter.convert(this::test_basic_types.parameters, test)
+        val deserializer = Deserializer()
+        val results = deserializer.convert(this::test_basic_types.parameters, test)
         assert(results[0] == "abc")
         assert(results[1] == false)
         assert(results[2] == 1.toShort())
@@ -50,8 +50,8 @@ class ConvertTests {
     fun test_dates(tdate: LocalDate, ttime: LocalTime, tlocaldatetime: LocalDateTime, tdatetime: DateTime):Unit{}
     @Test fun can_parse_dates(){
         val test = """{ "tdate": "2017-07-06", "ttime": "10:30:45", "tlocaldatetime": "2017-07-06T10:30:45", "tdatetime": "201707061030" }"""
-        val converter = Converter()
-        val results = converter.convert(this::test_dates.parameters, test)
+        val deserializer = Deserializer()
+        val results = deserializer.convert(this::test_dates.parameters, test)
         assert(results[0] == LocalDate.of(2017, 7, 6))
         assert(results[1] == LocalTime.of(10,30,45))
         assert(results[2] == LocalDateTime.of(2017,7,6, 10,30, 45))
@@ -68,8 +68,8 @@ class ConvertTests {
         val decDoub = MyEncryptor.encrypt("12345.67")
 
         val test = """{ "decString": "$decStr", "decInt": "$decInt", "decLong": "$decLong", "decDouble": "$decDoub" }"""
-        val converter = Converter(MyEncryptor)
-        val results = converter.convert(this::test_decrypted.parameters, test)
+        val deserializer = Deserializer(MyEncryptor)
+        val results = deserializer.convert(this::test_decrypted.parameters, test)
         assert((results[0] as EncString).value == "abc123")
         assert((results[1] as EncInt).value == 123)
         assert((results[2] as EncLong).value == 12345L)
@@ -80,8 +80,8 @@ class ConvertTests {
     fun test_arrays(strings: List<String>, bools:List<Boolean>, ints:List<Int>, longs:List<Long>, doubles:List<Double>):Unit {}
     @Test fun can_parse_arrays(){
         val test = """{ "strings": ["a", "b", "c"], "bools": [true, false, true], "ints": [1,2,3], "longs": [100,200,300], "doubles": [1.2,3.4,5.6] }"""
-        val converter = Converter(MyEncryptor)
-        val results = converter.convert(this::test_arrays.parameters, test)
+        val deserializer = Deserializer(MyEncryptor)
+        val results = deserializer.convert(this::test_arrays.parameters, test)
         assert((results[0] as List<String>)[0] == "a")
         assert((results[0] as List<String>)[1] == "b")
         assert((results[0] as List<String>)[2] == "c")
@@ -104,8 +104,8 @@ class ConvertTests {
     fun test_object(sample1: SampleObject1):Unit{}
     @Test fun can_parse_object(){
         val test = """{ "sample1": { "tstr": "abc", "tbool": false, "tshort": 1, "tint": 12, "tlong": 123, "tdoub": 123.45 } }"""
-        val converter = Converter()
-        val results = converter.convert(this::test_object.parameters, test)
+        val deserializer = Deserializer()
+        val results = deserializer.convert(this::test_object.parameters, test)
         assert(results[0] == ConvertTests.SampleObject1("abc", false, 1, 12, 123, 123.45))
     }
 
@@ -117,8 +117,8 @@ class ConvertTests {
             { "tstr": "abc", "tbool": false, "tshort": 1, "tint": 12, "tlong": 123, "tdoub": 123.45 },
             { "tstr": "def", "tbool": true , "tshort": 2, "tint": 34, "tlong": 456, "tdoub": 678.91 }
         ]}"""
-        val converter = Converter()
-        val inputs = converter.convert(this::test_object_list.parameters, test)
+        val deserializer = Deserializer()
+        val inputs = deserializer.convert(this::test_object_list.parameters, test)
         val results = inputs.get(0) as ArrayList<*>
         println(results)
         assert(results[0] == ConvertTests.SampleObject1("abc", false, 1, 12, 123, 123.45))
@@ -141,8 +141,8 @@ class ConvertTests {
                 ]
             }
         }"""
-        val converter = Converter()
-        val results = converter.convert(this::test_nested_object_list.parameters, test)
+        val deserializer = Deserializer()
+        val results = deserializer.convert(this::test_nested_object_list.parameters, test)
         val item = results[1] as NestedObject1
         assert(results[0] == "abc")
         assert(item.items[0] == ConvertTests.SampleObject1("abc", false, 1, 12, 123, 123.45))
@@ -155,8 +155,8 @@ class ConvertTests {
         val test = """{ "tstr": "abc", "tbool": false }"""
         val req = Request("a.b.c", listOf("a", "b", "c"), ApiConstants.SourceCLI, "post",
                 InputArgs(mapOf()), InputArgs(mapOf()))
-        val converter = Converter(null, mapOf("slatekit.common.Request" to { json, tpe -> req }))
-        val results = converter.convert(this::test_custom_converter.parameters, test)
+        val deserializer = Deserializer(null, mapOf("slatekit.common.Request" to { json, tpe -> req }))
+        val results = deserializer.convert(this::test_custom_converter.parameters, test)
         assert(results[0] == "abc")
         assert(results[1] == false)
         assert(results[2] == req)
